AgentAI/requirements.txt
requests>=2.31.0
pandas>=2.0.0
geopy>=2.4.0
pypdf>=3.17.0
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-experimental>=0.1.0
dataclasses-json>=0.6.0
nltk>=3.8
sympy>=1.12